* Suggested changes to move britepoolIdSystem out of userId default
