+++ modules/britepoolIdSystem.js
@@ -5,8 +5,9 @@
  * @requires module:modules/userId
  */
 
-import * as utils from '../../src/utils'
-import {ajax} from '../../src/ajax';
+import * as utils from '../src/utils'
+import {ajax} from '../src/ajax';
+import {submodule} from '../src/hook';
 
 /** @type {Submodule} */
 export const britepoolIdSubmodule = {
@@ -127,3 +128,5 @@ export const britepoolIdSubmodule = {
     return valueObj;
   }
 };
+
+submodule('userId', britepoolIdSubmodule);
