+++ modules/userId/index.js
@@ -106,7 +106,6 @@ import CONSTANTS from '../../src/constants.json';
 import {module} from '../../src/hook';
 import {unifiedIdSubmodule} from './unifiedIdSystem.js';
 import {pubCommonIdSubmodule} from './pubCommonIdSystem.js';
-import {britepoolIdSubmodule} from './britepoolIdSystem.js';
 
 const MODULE_NAME = 'User ID';
 const COOKIE = 'cookie';
@@ -551,6 +550,5 @@ init(config);
 // add submodules after init has been called
 attachIdSystem(pubCommonIdSubmodule);
 attachIdSystem(unifiedIdSubmodule);
-attachIdSystem(britepoolIdSubmodule);
 
 module('userId', attachIdSystem);
